+++ src/project/BlackJackPlayerActions.java
@@ -1,5 +1,7 @@
 package project;
 
+import java.util.ArrayList;
+import java.util.Iterator;
 import project.cards.Deck;
 
 /**
@@ -18,8 +20,10 @@ public class BlackJackPlayerActions extends PlayerActions {
 
     /**
      * Let the Player double down if they have enough money from
+     *
      * @param deck
      * @param user
+     * @throws java.lang.Exception
      */
     public static void doubleDown(Deck deck, BlackJackPlayer user) throws Exception {
         if (user.getCurrentBet() < user.getMoney()) {
@@ -29,4 +33,32 @@ public class BlackJackPlayerActions extends PlayerActions {
         stand(user);
     }
 
+    /**
+     * Get all BlackJackPlayers with the closest hand to 21
+     *
+     * @param users
+     * @return An ArrayList containing all the players with a hand that's value
+     * is 21 or the number closest to it
+     */
+    public static ArrayList<BlackJackPlayer> getWinners(ArrayList<BlackJackPlayer> users) {
+        ArrayList<BlackJackPlayer> winnerList = new ArrayList<>();
+        // First find the player with the closest value to 21
+        int closestValue = 0;
+        for (BlackJackPlayer currPlayer : users) {
+            if (calculateValue(currPlayer) <= 21 && calculateValue(currPlayer) >= closestValue) {
+                closestValue = calculateValue(currPlayer);
+                winnerList.add(currPlayer);
+            }
+        }
+        // Remove any Players who don't have the highest hand value
+        Iterator<BlackJackPlayer> itPlayer = winnerList.iterator();
+        while (itPlayer.hasNext()) {
+            BlackJackPlayer user = itPlayer.next();
+            if (calculateValue(user) < closestValue) {
+                itPlayer.remove();
+            }
+        }
+
+        return winnerList;
+    }
 }
